requirements.txt
# requirements.txt
numpy==1.21.6
pandas==1.4.3

apache-airflow==2.5.0

apache-airflow-providers-postgres==3.0.0

scikit-learn>=1.0
scipy>=1.7
joblib>=1.1

mlflow>=2.0

hydra-core>=1.2
omegaconf>=2.3

matplotlib>=3.4
seaborn>=0.11

pendulum==2.1.2
